# [libc++] Optimize the two range overload of mismatch